Change naming scheme
Instead of to-depspec/from-depspec, which is ambigious in that
a depspec can be a hash or a string, we now have depspec-hash()
and depspec-str(). Each of these may take a hash or a string, but
return the format implied by the routine name. This way you can
call depspec-str($str) to get back a normalized string depspec
without having to do depspec-str( depspec-hash($str) ).